Move ingest job status test code to core module
