+++ java/ingest/ingest-status-store/src/test/java/sleeper/ingest/status/store/testutils/DynamoDBIngestJobStatusStoreTestBase.java
@@ -57,13 +57,13 @@ import static sleeper.core.properties.table.TableProperty.TABLE_ID;
 import static sleeper.core.properties.table.TableProperty.TABLE_NAME;
 import static sleeper.core.record.process.status.ProcessStatusUpdateTestHelper.defaultUpdateTime;
 import static sleeper.core.statestore.AllReferencesToAFileTestHelper.filesWithReferences;
-import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.failedIngestJob;
-import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.finishedIngestJob;
-import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.finishedIngestRun;
-import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.ingestJobStatus;
-import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.ingestStartedStatus;
-import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.startedIngestJob;
-import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.startedIngestRun;
+import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.failedIngestJob;
+import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.finishedIngestJob;
+import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.finishedIngestRun;
+import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.ingestJobStatus;
+import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.ingestStartedStatus;
+import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.startedIngestJob;
+import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.startedIngestRun;
 import static sleeper.ingest.status.store.testutils.IngestStatusStoreTestUtils.createInstanceProperties;
 import static sleeper.ingest.status.store.testutils.IngestStatusStoreTestUtils.createSchema;
 import static sleeper.ingest.status.store.testutils.IngestStatusStoreTestUtils.createTableProperties;
